add option to fill blank spaces with provided value when composing cubes
when composing cubes, some combinations of dimension items will create empty spaces. By default, those spaces were being filled with the default value.
However, since now we only accept 0 or NaN as valid default values, cubes that need to be filled with 1s need that the consumer provide what the cube should be filled with.
This is a flag that we should re-evaluate having only 0 or Nan as valid default values.